Create README.md
I am a passionate and dynamic person, I like to cook and exercise with analytical skills and critical thinking. Academically, I have a degree in applied mathematics with skills in the areas of statistics, finance and computer science. Professionally, I have experience in business intelligence, user systems and university teaching.